Add feature flag for domain resolution input
